split rules
