+++ src/rules/mod.rs
@@ -1,0 +1,7 @@
+mod no_default_default;
+mod no_lazy_static;
+mod prefer_impl_param;
+
+pub use no_default_default::no_default_default_rule;
+pub use no_lazy_static::no_lazy_static_rule;
+pub use prefer_impl_param::prefer_impl_param_rule;
